+++ src/Arrays.php
@@ -30,7 +30,7 @@ final class Arrays
     /**
      * Checks if an array contains recursion.
      *
-     * @param mixed[] $arr The array to check.
+     * @param array<array-key, mixed> $arr The array to check.
      * @return bool True if the array contains recursion, false otherwise.
      */
     public static function containsRecursion(array $arr): bool